Update library versions for release Fix up signature related unit tests now that the latest release of xmlsec requires HMAC lengths >= 160

+++ import/java-xmltooling/branches/REL_1/src/test/java/org/opensaml/xml/signature/HMACSignatureTest.java
@@ -84,7 +84,7 @@ public class HMACSignatureTest extends XMLObjectBaseTestCase {
     protected void setUp() throws Exception {
         super.setUp();
         
-        hmacOutputLength = new Integer(64);
+        hmacOutputLength = new Integer(160);
         expectedKeyName = "KeyFoo123";
         
         SecretKey key = SecurityHelper.generateKey("AES", 128, null);
